requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
psycopg[binary]
alembic
pytest
python-jose[cryptography]
passlib[bcrypt]
python-multipart